Move helper into a different file
