+++ src/components/content/forms/conditions-autocomplete.tsx
@@ -1,43 +1,10 @@
 import { getAutoCompleteConditions } from "@/api/autocomplete-conditions";
 import { useCTW } from "@/components/core/ctw-provider";
-import {
-  SYSTEM_CONDITION_CLINICAL,
-  SYSTEM_CONDITION_VERIFICATION_STATUS,
-} from "@/fhir/system-urls";
-import { Condition } from "fhir/r4";
 import { InputHTMLAttributes, useState } from "react";
 import {
   ComboboxField,
   ComboxboxFieldOption,
 } from "../../core/form/combobox-field";
-
-// Sets any autofill values that apply when a user adds a condition, whether creating or confirming.
-export function setAddConditionDefaults(condition: Condition): void {
-  const addDefaults: Partial<Condition> = {
-    clinicalStatus: {
-      coding: [
-        {
-          system: SYSTEM_CONDITION_CLINICAL,
-          code: "active",
-          display: "Active",
-        },
-      ],
-      text: "active",
-    },
-    verificationStatus: {
-      coding: [
-        {
-          system: SYSTEM_CONDITION_VERIFICATION_STATUS,
-          code: "confirmed",
-          display: "Confirmed",
-        },
-      ],
-      text: "confirmed",
-    },
-  };
-
-  Object.assign(condition, addDefaults);
-}
 
 export type AutoCompleteComboboxProps = {
   defaultCoding?: fhir4.Coding;
